Upload images to AWS for reviews
