+++ server/src/routes/api/v1/reviewsRouter.js
@@ -1,7 +1,5 @@
 import express from 'express'
 import { Review } from '../../../models/index.js'
-import uploadImage from '../../../services/uploadImage.js'
-import objection from "objection"
 
 const reviewsRouter = new express.Router()
 
@@ -28,26 +26,6 @@ reviewsRouter.get("/image", async (req, res) => {
         const imageReview  = reviewToReturn.imageUrl 
         return res.status(200).json({ image: imageReview })
     } catch (error) {
-        return res.status(500).json({ errors: error })
-    }
-})
-
-reviewsRouter.post("/image", uploadImage.single("image"), async (req, res) => {
-    try {
-        const { body } = req
-        console.log(req.file.location)
-        console.log("after parse: ", body.json())
-
-        const data = {
-            ...body,
-            // image: req.file.location,
-            image: "",
-        }
-        const review = await Review.query().findById(req.review.id)
-        await req.review.query().insertAndFetch(data)
-        return res.status(201).json({ image: review.imageUrl })
-    } catch (error) {
-        console.log(error.message)
         return res.status(500).json({ errors: error })
     }
 })
